+++ out/pages/spriteeditor.js
@@ -79,6 +79,12 @@ class SpriteEditorPage {
                     case 'configuration':
                         this.saveConfiguration(message);
                         return;
+                    case 'loadPalette':
+                        this.loadPalette(message);
+                        return;
+                    case 'savePalette':
+                        this.savePalette(message);
+                        return;
                 }
                 // Unknown
                 console.log(`debugger:SpriteEditorPage: Unknown command called: ${message.command}`);
@@ -294,6 +300,117 @@ class SpriteEditorPage {
                     return false;
                 }
             }));
+            // Result
+            return true;
+        });
+    }
+    loadPalette(message) {
+        return __awaiter(this, void 0, void 0, function* () {
+            // Prompt user here, get selected file content
+            // and send response back to webview
+            // Prepare
+            let command = message.command;
+            //let content = message!.content;
+            //let file = message!.file;
+            // Options
+            let options = {
+                canSelectMany: false,
+                openLabel: "Open",
+                filters: {
+                    'Sprite Editor Palette': ['palette'],
+                    'All Files': ['*']
+                }
+            };
+            // Process
+            vscode.window.showOpenDialog(options).then((fileUri) => __awaiter(this, void 0, void 0, function* () {
+                if (fileUri && fileUri[0]) {
+                    // Process
+                    try {
+                        // Load
+                        let data = yield filesystem.ReadFileAsync(fileUri[0].fsPath);
+                        // Result
+                        this.currentPanel.webview.postMessage({
+                            command: command,
+                            status: 'ok',
+                            file: fileUri[0].fsPath,
+                            data: data
+                        });
+                    }
+                    catch (error) {
+                        // Result
+                        this.currentPanel.webview.postMessage({
+                            command: command,
+                            status: 'error',
+                            errorMessage: error
+                        });
+                        return false;
+                    }
+                }
+            }));
+            // Result
+            return true;
+        });
+    }
+    savePalette(message) {
+        return __awaiter(this, void 0, void 0, function* () {
+            // If no file provided open in workspace
+            // send response back to webview
+            // Prepare
+            let command = message.command;
+            let file = message.file;
+            let data = message.data;
+            let errorMessage = "";
+            // Set base uri
+            let fileUri = vscode.Uri.file(file);
+            // Prompt?
+            if (!file) {
+                // Options
+                let options = {
+                    defaultUri: vscode.Uri.file(filesystem.WorkspaceFolder()),
+                    saveLabel: "Save",
+                    filters: {
+                        'Sprite Editor Palette': ['palette'],
+                        'All Files': ['*']
+                    }
+                };
+                // Process
+                let result = yield vscode.window.showSaveDialog(options);
+                if (result)
+                    fileUri = result;
+            }
+            // Save?
+            if (fileUri) {
+                // Process
+                try {
+                    // Prepare
+                    let folder = path.dirname(fileUri.fsPath);
+                    // Save
+                    let result = yield filesystem.MkDirAsync(folder);
+                    if (result)
+                        result = yield filesystem.WriteFileAsync(fileUri.fsPath, data);
+                    // Validate
+                    if (result) {
+                        this.currentPanel.webview.postMessage({
+                            command: command,
+                            status: 'ok',
+                            file: fileUri.fsPath,
+                        });
+                        return true;
+                    }
+                    // Set
+                    errorMessage = "Failed to save palette";
+                }
+                catch (error) {
+                    errorMessage = error;
+                }
+                // Result
+                this.currentPanel.webview.postMessage({
+                    command: command,
+                    status: 'error',
+                    errorMessage: errorMessage
+                });
+                return false;
+            }
             // Result
             return true;
         });